+++ AnimThing.h
@@ -76,8 +76,8 @@ class CAnimThing : public CThing
 		int16_t	m_sLoop;								// Loops, if true.
 		char	m_szResName[RSP_MAX_PATH];		// Resource name.
 														
-		long	m_lAnimTime;						// Cummulative animation time.
-		long	m_lAnimPrevTime;					// Last animation time.
+		int32_t	m_lAnimTime;						// Cummulative animation time.
+		int32_t	m_lAnimPrevTime;					// Last animation time.
 														
 		U16			m_u16IdSendMsg;			// ID of CThing to send msg to when done.
 		GameMessage	m_msg;						// Message to send to m_pthingSendMsg.
