+++ Log.cpp
@@ -90,7 +90,7 @@ int16_t WriteTimeStamp(char *pszCaller,						// Name of calling routine
 							char *pszCalleeName,					// Name of player being sent or sending 
 							unsigned char ucMsgType,			// Message type
 							Net::SEQ seqStart,					// Beginning sequent sent/received
-							long sNum,								// Number of seq's sent/received
+							int32_t sNum,								// Number of seq's sent/received
 							char bReceived,							// a received or a sent message? TRUE if received
 							U16 u16PackageID/*=0*/)				// Uniquely identifiable package id																		//		True if receiving, false if sending
 	{	
@@ -99,7 +99,7 @@ int16_t WriteTimeStamp(char *pszCaller,						// Name of calling routine
 	char szTime[256]; 
 	char szSeq[256];
 	char szNum[256];
-	long lTime = rspGetMilliseconds();
+	int32_t lTime = rspGetMilliseconds();
 
 	if ((ucMsgType == NetMsg::START_REALM)&&(bReceived))
 		{
@@ -148,7 +148,7 @@ int16_t WriteTimeStamp(char *pszCaller,						// Name of calling routine
 
 	// Write package ID
 	char szPackageID[256];
-	ltoa((long)u16PackageID, szPackageID, 10);
+	ltoa((int32_t)u16PackageID, szPackageID, 10);
 	prfLog->Write(szPackageID);
 	prfLog->Write(" ");
